FBV blog 초안 완성

+++ blog/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect, get_object_or_404 
+from django.contrib.auth.decorators import login_required
 from django.core.paginator import Paginator
 from .models import Post, Category
 from .forms import PostForm
@@ -29,7 +30,7 @@ def my_view(request):
     }
     return render(request, 'my_template.html', data)
 
-
+@login_required
 def post_create(request):
     if request.method == "POST": # Post요청 확인
         form = PostForm(request.POST) # PostFrom 인스턴스 생성 인자 전달
@@ -42,24 +43,50 @@ def post_create(request):
         form = PostForm()
     return render(request, 'blog/post_form.html', {'form': form})
 
+# @login_required
+# def post_edit(request, pk):
+#     post = get_object_or_404(Post, pk=pk)
+#     if request.method == "POST":
+#         form = PostForm(request.POST, instance=post) # 이미 데이터베이스에 저장된 게시물을 수정 instance = post 기존 Post객체를 폼에 제공
+#         if post.author != request.user:
+#             return redirect('post_detail', pk=pk)
+    
+#         if form.is_valid():
+#             post = form.save(commit=False)
+#             post.author = request.user  # 현재는 로그인 기능이 없으므로 나중에 구현
+#             post.save()
+#             return redirect('post_detail', pk=post.pk)
+#     else:
+#         form = PostForm(instance=post)
+#     return render(request, 'blog/post_form.html', {'form': form})
+
+@login_required
 def post_edit(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    
     if request.method == "POST":
-        form = PostForm(request.POST, instance=post) # 이미 데이터베이스에 저장된 게시물을 수정 instance = post 기존 Post객체를 폼에 제공
+        form = PostForm(request.POST, instance=post)
         if form.is_valid():
             post = form.save(commit=False)
-            post.author = request.user  # 현재는 로그인 기능이 없으므로 나중에 구현
+            if post.author != request.user:
+                return redirect('post_detail', pk=pk)  # 작성자만 수정 가능
             post.save()
             return redirect('post_detail', pk=post.pk)
     else:
         form = PostForm(instance=post)
     return render(request, 'blog/post_form.html', {'form': form})
 
+@login_required
 def post_delete(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    if post.author != request.user:
+        return redirect('post_detail', pk=pk)
     if request.method == "POST":
         post.delete()
         return redirect('post_list')
     return render(request, 'blog/post_confirm_delete.html', {'post': post})
 
-
+def category_posts(request, category_slug):
+    category = get_object_or_404(Category, slug=category_slug)
+    posts = Post.objects.filter(category=category).order_by('-created_date')
+    return render(request, 'blog/category_posts.html', {'category': category, 'posts': posts})
